break out the sparse checkout functionality
